requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
djangorestframework==3.11.0
idna==2.9
Pillow==7.1.1
PyPDF2==1.26.0
pytz==2019.3
requests==2.23.0
selenium==3.141.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
